Convert tests to use RuleTester instead of --report-unused-disable-directives (#31)
**Public-Facing Changes**
- `no-boolean-parameters` correctly reports the function name in variable assignments like `const x = (a: boolean) => {}`
- `no-meaningless-void-operator` no longer reports errors on `never` values by default; added `checkNever` option to enable (imported these changes from https://github.com/typescript-eslint/typescript-eslint/pull/3641)

**Description**
The `RuleTester` provides the ability to assert on specific error messages and fixes that rules produce. It's also more in line with how the upstream projects (eslint and typescript-eslint) test their rules, which will smooth the process if we decide to upstream these rules.

The tests for (and changes to) no-meaningless-void-operator are changes I made based on feedback on my typescript-eslint PR: https://github.com/typescript-eslint/typescript-eslint/pull/3641

Imported jest.config.js from template-typescript.