Making header rule time dynamic (#9)
Altering the header rule so that header that is present on Lichtblick,
and other repositories, is always updated regarding the year we're on.
Also made it that if a file already has an header but with the wrong
year it will replace with the correct header.

NOTE: Everytime the year changes the first PR will trigger lint errors on all files with an header.